requirements.txt
streamlit
pandas
pymongo
python-dotenv
certifi